Export ribbon code as a common function, localize environment strings
